Move event interpretation functionality to its own class.
