+++ HouseRules_Core/HouseRules_Core.csproj
@@ -47,6 +47,7 @@
         </Reference>
     </ItemGroup>
     <ItemGroup>
+        <Compile Include="EventInterpreter.cs" />
         <Compile Include="ModPatcher.cs" />
         <Compile Include="Properties\AssemblyInfo.cs" />
         <Compile Include="HR.cs" />
