+++ HouseRules_Core/ModPatcher.cs
@@ -2,16 +2,15 @@
 {
     using System.Reflection;
     using Boardgame;
-    using Boardgame.BoardEntities;
     using Boardgame.Networking;
     using Boardgame.SerializableEvents;
-    using DataKeys;
     using HarmonyLib;
 
     internal class ModPatcher
     {
         private static GameContext _gameContext;
         private static bool _isStartingGame;
+        private static EventInterpreter _eventInterpreter;
 
         internal static void Patch(Harmony harmony)
         {
@@ -56,6 +55,7 @@
         {
             var gameContext = Traverse.Create(__instance).Field<GameContext>("gameContext").Value;
             _gameContext = gameContext;
+            _eventInterpreter = EventInterpreter.NewInstance(_gameContext);
         }
 
         private static void CreatingGameState_OnJoinedRoom_Prefix()
@@ -120,82 +120,12 @@
                 return;
             }
 
-            if (!DoesEventRepresentNewSpawn(serializableEvent))
+            if (!_eventInterpreter.DoesEventRepresentNewSpawn(serializableEvent))
             {
                 return;
             }
 
             _gameContext.serializableEventQueue.SendResponseEvent(SerializableEvent.CreateRecovery());
-        }
-
-        private static bool DoesEventRepresentNewSpawn(SerializableEvent serializableEvent)
-        {
-            switch (serializableEvent.type)
-            {
-                case SerializableEvent.Type.SpawnPiece:
-                case SerializableEvent.Type.UpdateFogAndSpawn:
-                case SerializableEvent.Type.SetBoardPieceID:
-                case SerializableEvent.Type.SlimeFusion:
-                case SerializableEvent.Type.GoToNextLevel:
-                    return true;
-            }
-
-            if (serializableEvent.type == SerializableEvent.Type.OnAbilityUsed)
-            {
-                return DoesAbilityEventRepresentNewSpawn((SerializableEventOnAbilityUsed)serializableEvent);
-            }
-
-            if (serializableEvent.type == SerializableEvent.Type.PieceDied)
-            {
-                return DoesPieceDiedEventRepresentNewSpawn((SerializableEventPieceDied)serializableEvent);
-            }
-
-            return false;
-        }
-
-        private static bool DoesAbilityEventRepresentNewSpawn(SerializableEventOnAbilityUsed onAbilityUsedEvent)
-        {
-            var abilityKey = Traverse.Create(onAbilityUsedEvent).Field<AbilityKey>("abilityKey").Value;
-            switch (abilityKey)
-            {
-                case AbilityKey.SummonElemental:
-                case AbilityKey.SummonBossMinions:
-                case AbilityKey.NaturesCall:
-                case AbilityKey.Tornado:
-                case AbilityKey.MonsterBait:
-                case AbilityKey.ProximityMine:
-                case AbilityKey.EyeOfAvalon:
-                case AbilityKey.SwordOfAvalon:
-                case AbilityKey.BeaconOfSmite:
-                case AbilityKey.BeaconOfHealing:
-                case AbilityKey.RaiseRoots:
-                case AbilityKey.CallCompanion:
-                    return true;
-            }
-
-            var abilityName = abilityKey.ToString();
-            var isSpawnAbility = abilityName.Contains("Spawn");
-            var isLampAbility = abilityName.Contains("Lamp");
-
-            return isSpawnAbility || isLampAbility;
-        }
-
-        private static bool DoesPieceDiedEventRepresentNewSpawn(SerializableEventPieceDied pieceDiedEvent)
-        {
-            foreach (var pieceId in pieceDiedEvent.deadPieces)
-            {
-                if (!_gameContext.pieceAndTurnController.TryGetPiece(pieceId, out Piece piece))
-                {
-                    continue;
-                }
-
-                if (piece.boardPieceId == BoardPieceId.SpiderEgg)
-                {
-                    return true;
-                }
-            }
-
-            return false;
         }
     }
 }
